external_agent/requirements.txt
a2a-sdk[http-server]>=0.2.5
uvicorn>=0.27.0
httpx>=0.27.0
openai>=1.0.0
